+++ src/phishbench/feature_extraction/url/features/__init__.py
@@ -2,6 +2,7 @@
 This module contains the built-in URL/website features
 """
 from .raw_url_features import *
+from .url_token_features import *
+from .url_whitelist_features import *
 from .website_features import *
 from .website_html_features import *
-from .url_whitelist_features import *