+++ tests/unit_tests/features/url/test_url_features.py
@@ -1,10 +1,10 @@
+"""
+Tests for raw url features
+"""
 import unittest
-from unittest.mock import patch
 
 import phishbench.feature_extraction.url.features as url_features
-from phishbench import Features
 from phishbench.input import URLData
-from .... import mock_objects
 
 
 # pylint: disable=missing-function-docstring
@@ -234,70 +234,73 @@ class TestURLReflectionFeatures(unittest.TestCase):
         result = url_features.double_slashes_in_path(test_url)
         self.assertEqual(2, result)
 
+    def test_has_www_in_middle_false(self):
+        test_url = URLData('http://www.google.com', download_url=False)
+        result = url_features.has_www_in_middle(test_url)
+        self.assertFalse(result)
 
-@patch('phishbench.utils.phishbench_globals.config', new_callable=mock_objects.get_mock_config)
-class TestURLFeatures(unittest.TestCase):
-    # pylint: disable=invalid-name
-    def test_kolmogorov_shmirnov(self, config_mock):
-        pass
+    def test_has_www_in_middle_domain(self):
+        test_url = URLData('http://www.google.com.www.test.com', download_url=False)
+        result = url_features.has_www_in_middle(test_url)
+        self.assertTrue(result)
 
-    def test_URL_Kullback_Leibler_Divergence(self, config_mock):
-        pass
+    def test_has_www_in_middle_path(self):
+        test_url = URLData('http://www.google.com/www.test.com', download_url=False)
+        result = url_features.has_www_in_middle(test_url)
+        self.assertTrue(result)
 
-    def test_URL_english_frequency_distance(self, config_mock):
-        pass
+    def test_port_protocol_match_no_port(self):
+        test_url = URLData('http://www.google.com/www.test.com', download_url=False)
+        result = url_features.protocol_port_match(test_url)
+        self.assertTrue(result)
 
-    def test_URL_Token_Count(self, config_mock):
-        config_mock['URL_Features']['Token_Count'] = "True"
-        list_features = {}
-        list_time = {}
+    def test_port_protocol_match(self):
+        test_url = URLData('https://www.google.com:443/abc', download_url=False)
+        result = url_features.protocol_port_match(test_url)
+        self.assertTrue(result)
 
-        Features.URL_Token_Count('http://te2t-url.com/home.html', list_features, list_time)
+    def test_port_protocol_match_false(self):
+        test_url = URLData('https://www.google.com:8080/abc', download_url=False)
+        result = url_features.protocol_port_match(test_url)
+        self.assertFalse(result)
 
-        self.assertEqual(6, list_features["Token_Count"], 'incorrect Token_Count')
+    def test_port_protocol_match_unknown(self):
+        test_url = URLData('stratum+tcp://scrypt.LOCATION.nicehash.com:3333', download_url=False)
+        result = url_features.protocol_port_match(test_url)
+        self.assertFalse(result)
 
-    def test_URL_Average_Path_Token_Length(self, config_mock):
-        config_mock['URL_Features']['Average_Path_Token_Length'] = "True"
-        list_features = {}
-        list_time = {}
 
-        Features.URL_Average_Path_Token_Length('http://te2t-url.com/home.html', list_features, list_time)
+class TestURLTokenFeatures(unittest.TestCase):
+    """
+    Tests url_token_features
+    """
 
-        self.assertEqual(list_features["Average_Path_Token_Length"], 8 / 3, 'incorrect Average_Path_Token_Length')
+    def test_token_count(self):
+        test_url = URLData('http://te2t-url.com/home.html', download_url=False)
+        result = url_features.token_count(test_url)
 
-    def test_URL_Average_Domain_Token_Length(self, config_mock):
-        config_mock['URL_Features']['Average_Domain_Token_Length'] = "True"
-        list_features = {}
-        list_time = {}
+        self.assertEqual(6, result)
 
-        Features.URL_Average_Domain_Token_Length('http://te2t-url.com/home.html', list_features, list_time)
+    def test_average_path_token_length(self):
+        test_url = URLData('http://te2t-url.com/home.html', download_url=False)
+        result = url_features.average_path_token_length(test_url)
 
-        self.assertEqual(list_features["Average_Domain_Token_Length"], 10 / 3, 'incorrect Average_Domain_Token_Length')
+        self.assertEqual(4, result)
 
-    def test_URL_Longest_Domain_Token(self, config_mock):
-        config_mock['URL_Features']['Longest_Domain_Token'] = "True"
-        list_features = {}
-        list_time = {}
+    def test_average_path_token_length_no_path(self):
+        test_url = URLData('http://te2t-url.com', download_url=False)
+        result = url_features.average_path_token_length(test_url)
 
-        Features.URL_Longest_Domain_Token('http://te2t-url.com/home.html', list_features, list_time)
+        self.assertEqual(0, result)
 
-        self.assertEqual(4, list_features["Longest_Domain_Token"], 'incorrect Longest_Domain_Token')
+    def test_average_domain_token_length(self):
+        test_url = URLData('http://te2t-url.com/home.html', download_url=False)
+        result = url_features.average_domain_token_length(test_url)
 
-    def test_URL_Protocol_Port_Match(self, config_mock):
-        config_mock['URL_Features']['Protocol_Port_Match'] = "True"
-        list_features = {}
-        list_time = {}
+        self.assertEqual(10 / 3, result)
 
-        Features.URL_Protocol_Port_Match('http://te2t-url.com/home.html', list_features, list_time)
+    def test_longest_domain_token_length(self):
+        test_url = URLData('http://te2t-url.com/home.html', download_url=False)
+        result = url_features.longest_domain_token_length(test_url)
 
-        self.assertEqual(list_features["Protocol_Port_Match"], 1, 'incorrect Protocol_Port_Match')
-
-    def test_URL_Has_www_in_Middle(self, config_mock):
-        config_mock['URL_Features']['Has_WWW_in_Middle'] = "True"
-        test_url = "http://httpwwwchase.com"
-        list_features = {}
-        list_time = {}
-
-        Features.URL_Has_WWW_in_Middle(test_url, list_features, list_time)
-
-        self.assertEqual(list_features["Has_WWW_in_Middle"], 1, 'incorrect Has_WWW_in_Middle')
+        self.assertEqual(4, result)